app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "ttkbootstrap>=1.12.1",
]
